Introduce the Spring AMQP to handle the events
##Description
This application will be transformed from a simple monolithic application to a microservice. 
Spring AMQP (with RABBITMQ) will be used to handle event messaging bus. 

In this commit we added : 
- A RabbitMQ configuration class to use events in our application, we changed the serialisation type from Java Object to JSON 
- an event dispatcher to handle the communication with the event bus 
- an event to model the multiplication attempt. 

After each attempt, and in the transactional method, the result of the multiplication will be modeled to an event and sent to the Event Bus after the save in databse. To do it, MultiplicatonServiceImpl and its test class have been changed.


## Tests 

Execute all the unit tests

## Database 
Nil